chore: refactor package structure
