+++ pkg/cli/completion_test.go
