+++ pkg/cli/config_test.go
@@ -25,9 +25,10 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/tbckr/sgpt/v2/pkg/api"
+
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
-	"github.com/tbckr/sgpt/v2/api"
 )
 
 func TestConfigCmd(t *testing.T) {
