+++ pkg/cli/licenses.go
