+++ pkg/cli/man.go
