+++ pkg/cli/man_test.go
