+++ pkg/shell/shell_test.go
